fix(protocol): make native engine telemetry parseable
`ExecutionEngineTelemetryV2::Native` declared `compile_nanos` and
`run_nanos` as `u128`. Serde buffers an internally tagged enum's content
through `Content`, which has no 128-bit carrier, so the variant could be
constructed and serialized and then never read back: no consumer could
parse a native execution report at all.

Both counters are `u64` nanoseconds now — 584 years, so nothing
representable is lost — and the VM narrows its internal `u128`
accumulators at the producer with a saturating conversion, so the SDK
report type and the protocol summary agree on one width.

The CLI's projection function stays: the VM's native report carries the
full diagnostic counter set while the protocol variant denies unknown
fields, which is a separate reason for the narrowing step. What changed
is that what it prints is now parseable, and a CLI test asserts exactly
that by feeding `rss run --trusted-in-process --native --json` output
back through `serde_json::from_str::<ExecutionReportV2>`.

The JSON is unchanged for every value below 2^64 and the checked-in v2
schema never constrained the width, so there is no migration; ADR 0235
records why none exists.

+++ crates/rsscript-sdk/tests/native_jit_scorecard.rs
@@ -545,7 +545,7 @@ fn native_jit_pass_scorecard() {
                 "cold_e2e_native_mad_ns": native_mad.as_nanos(),
                 "warm_native_instrumented_ns": diagnostic_native_run_nanos,
                 "instrumented_native_nanos_per_entry": diagnostic_native_run_nanos
-                    / u128::from(native_calls.saturating_add(osr_entries).saturating_add(continuation_entries).max(1)),
+                    / native_calls.saturating_add(osr_entries).saturating_add(continuation_entries).max(1),
                 "speedup": interpreter.as_secs_f64() / native.as_secs_f64(),
                 "translation_nanos": translation_nanos,
                 "validation_nanos": validation_nanos,
